fail on compiler warnings and fix some
fix several warnings emitted from the compiler: unused imports, numeric
widening, dead code etc.

+++ src/main/scala/chee/properties/Converter.scala
@@ -25,7 +25,7 @@ object MegaPixelConverter extends Converter[Int] {
   def parse(str: String) =
     megaPixel.parse(str).right.map(_.toInt)
 
-  override def render(p: Int) = megaPixel.render(p)
+  override def render(p: Int) = megaPixel.render(p.toLong)
 }
 
 object FileSizeConverter extends Converter[Long] {
